# SY_project02/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "My API",
    "description": "Description",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/payment/payment/success": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "paymentKey",
            "in": "query",
            "type": "string"
          },
          {
            "name": "orderId",
            "in": "query",
            "type": "string"
          },
          {
            "name": "amount",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/payment/payment/fail": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "message",
            "in": "query",
            "type": "string"
          },
          {
            "name": "code",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
...
  }
}